Factor annotated test handling into a TestUtils module.
