+++ Sources/TestUtils/AnnotatedValFileTest.swift
@@ -17,6 +17,32 @@ extension Diagnostic {
 
 extension XCTestCase {
 
+  /// Returns the result of invoking `f` on an initially-empty diagnostic set, reporting any
+  /// diagnostics added and/or thrown as XCTest issues.
+  public func checkNoDiagnostic<R>(
+    f: (inout DiagnosticSet) throws -> R, testFile: StaticString = #filePath, line: UInt = #line
+  ) rethrows -> R {
+    var d = DiagnosticSet()
+    do {
+      let r = try f(&d)
+      checkEmpty(d)
+      return r
+    } catch let d1 as DiagnosticSet {
+      XCTAssertEqual(
+        d1, d, "thrown diagnostics don't match mutated diagnostics",
+        file: testFile, line: line)
+      checkEmpty(d)
+      throw d
+    }
+  }
+
+  /// Reports any diagnostics in `s` as XCTest issues.
+  public func checkEmpty(_ s: DiagnosticSet) {
+    for d in s.elements {
+      record(XCTIssue(.error("unexpected diagnostic: '\(d.message)'", at: d.site, notes: d.notes)))
+    }
+  }
+
   /// The effects of running the `processAndCheck` parameter to `checkAnnotatedValFiles`.
   fileprivate typealias ProcessingEffects = (
     /// Whether the processing completed without Val errors.
@@ -32,30 +58,6 @@ extension XCTestCase {
     .deletingLastPathComponent()
     .deletingLastPathComponent()
 
-  /// Applies `process` to each ".val" file in the test suite at `suitePath` (relative to the
-  /// `ValTests/` directory of this project) and reports XCTest failures where the effects of
-  /// processing don't match the file's annotation commands ("diagnostic", "expect-failure", and
-  /// "expect-success").
-  ///
-  /// - Parameter process: applies some compilation phases to `file`, updating `diagnostics`
-  ///   with any generated diagnostics. Throws an `Error` if any phases failed.
-  func checkAnnotatedValFileDiagnostics(
-    inSuiteAt suitePath: String,
-    _ process: (_ file: SourceFile, _ diagnostics: inout DiagnosticSet) throws -> Void
-  ) throws {
-    let annotatedValFiles = try testSuite(at: suitePath)
-    for f in annotatedValFiles {
-      try checkAnnotations(
-        in: f, checkingAnnotationCommands: [],
-        { (file, annotationsToHandle, diagnostics) in
-          assert(annotationsToHandle.isEmpty)
-          try process(file, &diagnostics)
-          return []
-        }
-      )
-    }
-  }
-
   /// Applies `processAndCheck` to `valToTest`, along with the subset of its annotations
   /// whose commands match `checkedCommands`, and reports resulting XCTest failures, along with any
   /// additional failures where the effects of processing don't match the its annotation
@@ -66,7 +68,7 @@ extension XCTestCase {
   ///   - processAndCheck: applies some compilation phases to `file`, updating `diagnostics`
   ///     with any generated diagnostics, then checks `annotationsToCheck` against the results,
   ///     returning corresponding test failures. Throws an `Error` if any phases failed.
-  func checkAnnotations(
+  public func checkAnnotations(
     in valToTest: SourceFile,
     checkingAnnotationCommands checkedCommands: Set<String> = [],
     _ processAndCheck: (
@@ -100,24 +102,6 @@ extension XCTestCase {
 
     for f in failures {
       record(f)
-    }
-  }
-
-  /// Returns the Val test cases in the suite at `path` relative to the `ValTests/` directory of
-  /// this project.
-  ///
-  /// The suite is sought at `path` relative to `ValTests/`. If no such directory exists, the
-  /// suite is sought at `path` relative to the root of the resource bundle associated with the
-  /// current Swift module.
-  fileprivate func testSuite(
-    at path: String
-  ) throws -> [SourceFile] {
-    let p = URL(fileURLWithPath: path, relativeTo: XCTestCase.valTests)
-    if p.hasDirectoryPath {
-      return try sourceFiles(in: [p])
-    } else {
-      let s = Bundle.module.url(forResource: path, withExtension: nil)!
-      return try sourceFiles(in: [s])
     }
   }
 
@@ -160,6 +144,28 @@ extension XCTestCase {
       XCTIssue(.error("unexpected diagnostic: '\($0.message)'", at: $0.site, notes: $0.notes))
     }
     return testFailures
+  }
+
+}
+
+extension XCTIssue {
+
+  /// Creates an instance from a diagnostic.
+  init(_ d: Diagnostic) {
+    self.init(
+      type: .assertionFailure,
+      compactDescription: d.message,
+      sourceCodeContext:
+        .init(location: XCTSourceCodeLocation.init(d.site.first())))
+  }
+
+}
+
+extension XCTSourceCodeLocation {
+
+  /// Creates an instance from a location in a Val source file.
+  convenience init(_ l: SourcePosition) {
+    self.init(fileURL: l.file.url, lineNumber: l.line.number)
   }
 
 }
