+++ Tests/EndToEndTests/ExecutionTests.swift
@@ -1,5 +1,6 @@
 import ArgumentParser
 import Core
+import TestUtils
 import ValCommand
 import XCTest
 
@@ -19,6 +20,32 @@ final class ExecutionTests: EndToEndTestCase {
 }
 
 class EndToEndTestCase: XCTestCase {
+
+  /// Compiles and runs the given file at `valFilePath`, `XCTAssert`ing that diagnostics and exit
+  /// codes match annotated expectations.
+  func compileAndRun(_ valFilePath: String, file: StaticString = #filePath, line: UInt = #line)
+    throws
+  {
+    #if false
+      checkAnnotatedValFileDiagnostics(inFileAt: valFilePath) {
+        (_ file: SourceFile, _ diagnostics: inout DiagnosticSet) in
+        let output = try compile(f, with: ["--emit", "binary"])
+      }
+    #endif
+    let f = URL(fileURLWithPath: valFilePath)
+    let output = try compile(f, with: ["--emit", "binary"])
+    do {
+      let (status, _) = try run(output)
+      XCTAssertEqual(
+        status, 0,
+        "Execution of binary for test \(f.lastPathComponent) failed with exit code \(status)",
+        file: file, line: line
+      )
+    } catch {
+      XCTFail(
+        "While testing \(f.lastPathComponent), cannot execute: \(output)", file: file, line: line)
+    }
+  }
 
   /// Compiles `input` with the given arguments and returns the URL of the output file.
   ///
